requirements.txt
kafka
paho-mqtt
psycopg2
boto3
pymongo==3.7.1
